When Back to my original dgGoolgel class for Big number
for some reason the implemnention based on 
// based of the paper of: Jonathan Richard Shewchuk October 1, 1997
// "Adaptive Precision Floating-Point Arithmetic and Fast Robust Geometric Predicates"

is very erratic and is is not producing the desires results, basically it doe all the work 
for no real gain, and this is generation many, many round off error is algorithm that were working perfectly before.

I had made many modification not knwo that me problem started when I change my call for Jonathan Richard Shewchuk
Theorem on exact arithmetic Float.

I am sure his theory s correct, but I suspect the reason I do no get it to work is that architectures like Intel
float points register has more intermediate bits that the memory representation and truncation error
make his theorem breaks down because of round of error.

Me implementation base of Integers, is more trivial be its more reliable on all architectures.

I also revamped the old implementation based on 32 bit integers and 32 bit float
to work with 64 bit integers and 64 double, since most CPUs today are 64 bit possessors with 
64 bit integer arithmetic and fast double precision floats.

I in the future we need to move to a Console with 32 bit support only, the call can essay be modified again





 
 
 

 




git-svn-id: http://newton-dynamics.googlecode.com/svn@601 b7a2f1d6-d59d-a8fe-1e9e-8d4888b32692